Issue #274 Added parameterization of preexisting file
Per a comment on the pull request, I added parameterization of a
pre-existing file

+++ tests/test_file_handler.py
@@ -170,15 +170,17 @@ def test_timed_rotating_file_handler__rollover_format(tmpdir, activation_strateg
 
 
 @pytest.mark.parametrize("backup_count", [1, 3])
+@pytest.mark.parametrize("preexisting_file", [True, False])
 def test_timed_rotating_file_handler__not_timed_filename_for_current(
-        tmpdir, activation_strategy, backup_count
+        tmpdir, activation_strategy, backup_count, preexisting_file
 ):
     basename = str(tmpdir.join('trot.log'))
 
-    with open(basename, 'w') as file:
-        file.write('contents')
-    jan_first = time.mktime(datetime(2010, 1, 1).timetuple())
-    os.utime(basename, (jan_first, jan_first))
+    if preexisting_file:
+        with open(basename, 'w') as file:
+            file.write('contents')
+        jan_first = time.mktime(datetime(2010, 1, 1).timetuple())
+        os.utime(basename, (jan_first, jan_first))
 
     handler = logbook.TimedRotatingFileHandler(
         basename,
@@ -205,17 +207,14 @@ def test_timed_rotating_file_handler__not_timed_filename_for_current(
         for x in xrange(20):
             handler.handle(fake_record('Last One', 2010, 1, 8, x + 1))
 
-    computed_files = sorted(
-        x for x in os.listdir(str(tmpdir)) if x.startswith('trot')
-    )
+    computed_files = [x for x in os.listdir(str(tmpdir)) if x.startswith('trot')]
 
-    expected_files = sorted((
-        ['trot.log.2010-01-01'] +
-        ['trot.log.2010-01-0{0}'.format(i) for i in xrange(5, 8)] +
-        ['trot.log']
-    )[-backup_count:])
+    expected_files = ['trot.log.2010-01-01'] if preexisting_file else []
+    expected_files += ['trot.log.2010-01-0{0}'.format(i) for i in xrange(5, 8)]
+    expected_files += ['trot.log']
+    expected_files = expected_files[-backup_count:]
 
-    assert computed_files == expected_files
+    assert sorted(computed_files) == sorted(expected_files)
 
     with open(str(tmpdir.join('trot.log'))) as f:
         assert f.readline().rstrip() == '[01:00] Last One'
